+++ src/app.module.ts
@@ -2,9 +2,11 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { HealthModule } from './modules/health/health.module';
+import { AuthModule } from './modules/auth/auth.module';
+import { PrismaModule } from './prisma/prisma.module';
 
 @Module({
-  imports: [HealthModule],
+  imports: [PrismaModule, HealthModule, AuthModule],
   controllers: [AppController],
   providers: [AppService],
 })